Seeding implementation, login style and database tables modified
I modified the name of the table list for activity and added the table category, also i modified the name of the module to_do for activity. In the seeding I defined a user and three basic categories, also I unable the loggin of the user in the file auth.py, in a future i am going to declare a field in the database to determine if the user is logable or not. I separate the login and register page from the base template and implemented part of the style to the login page